Test finalized
The answers for the first part are on the txt file at the root.

For the code test I have created a React app for the front-end (front folder). I have provided both the source-code and the build (front/build). To run it just open the index.html in the front/build folder in the browser.

For the back-end part I've created an API (api folder) with the Lumen framework by Laravel. I've used composer with PHP7.4 to run a local server with the command line(in the api folder):

php -S localhost:8000 -t public

Please rename the api/.env.example to .env and change the database connection information with the proper ones for your local envirornment.

If there's any problem please contact me.